Also do so for implementations
